feat: memory.html 导入/导出入口——聊天记录导入 + 记忆库导出
- services/chatParser.js + chatImport.js: 解析/写入逻辑抽成共享模块（CLI 与 API 复用）
- routes/import.js: POST /api/import/chat + GET /api/export/memory
- scripts/import_chat.js: 复用共享模块
- memory.html: 顶栏「导入/导出」按钮 + 导入弹窗（拖拽/预览/确认/结果），导出直接下载 JSONL
